+++ python/examples/scenarios/collector_scenario.py
@@ -7,6 +7,7 @@ import numpy as np
 from google.protobuf.json_format import MessageToDict
 from protobuf_APIs import falconeye_pb2, radar_pb2
 
+from scenarios.meta.collector_meta import get_builder
 from scenarios.utils.com_manager import ComManager, MqttConst
 from scenarios.utils.filesystem import get_collector_instances, load_config
 from scenarios.utils.gis import transform_combine_to_local, utm_array_to_local, get_combine_region, latlon_to_utm
@@ -18,7 +19,6 @@ from scenarios.safety_subsystems.radar_filter import RadarFilter
 from scenarios.safety_subsystems.path_prediction import get_path_prediction
 
 import xviz
-import xviz.builder as xbuilder
 
 TRACTOR_GPS_TO_REAR_AXLE = 1.9304
 COMBINE_GPS_TO_CENTER = 1.0
@@ -95,133 +95,7 @@ class CollectorScenario:
 
     def get_metadata(self):
         if not self._metadata:
-            builder = xviz.XVIZMetadataBuilder()
-            builder.stream("/vehicle_pose")\
-                .category(xviz.CATEGORY.POSE)
-
-            builder.stream("/radar_filtered_out_targets")\
-                .coordinate(xviz.COORDINATE_TYPES.VEHICLE_RELATIVE)\
-                .stream_style({'fill_color': [255, 255, 0]})\
-                .category(xviz.CATEGORY.PRIMITIVE)\
-                .type(xviz.PRIMITIVE_TYPES.CIRCLE)
-            builder.stream("/radar_passed_filter_targets")\
-                .coordinate(xviz.COORDINATE_TYPES.VEHICLE_RELATIVE)\
-                .stream_style({'fill_color': [255, 0, 0]})\
-                .category(xviz.CATEGORY.PRIMITIVE)\
-                .type(xviz.PRIMITIVE_TYPES.CIRCLE)
-            builder.stream("/radar_id")\
-                .coordinate(xviz.COORDINATE_TYPES.VEHICLE_RELATIVE)\
-                .stream_style({'fill_color': [0, 0, 0]})\
-                .category(xviz.CATEGORY.PRIMITIVE)\
-                .type(xviz.PRIMITIVE_TYPES.TEXT)
-
-            builder.stream("/camera_targets")\
-                .coordinate(xviz.COORDINATE_TYPES.VEHICLE_RELATIVE)\
-                .stream_style({'fill_color': [0, 255, 255]})\
-                .category(xviz.CATEGORY.PRIMITIVE)\
-                .type(xviz.PRIMITIVE_TYPES.CIRCLE)
-
-            builder.stream("/tracking_targets")\
-                .coordinate(xviz.COORDINATE_TYPES.VEHICLE_RELATIVE)\
-                .category(xviz.CATEGORY.PRIMITIVE)\
-                .type(xviz.PRIMITIVE_TYPES.CIRCLE)
-            builder.stream("/tracking_id")\
-                .coordinate(xviz.COORDINATE_TYPES.VEHICLE_RELATIVE)\
-                .stream_style({'fill_color': [0, 0, 0]})\
-                .category(xviz.CATEGORY.PRIMITIVE)\
-                .type(xviz.PRIMITIVE_TYPES.TEXT)
-
-            combine_color = [128, 0, 128]
-            builder.stream("/combine_position")\
-                .coordinate(xviz.COORDINATE_TYPES.VEHICLE_RELATIVE)\
-                .stream_style({'fill_color': combine_color})\
-                .category(xviz.CATEGORY.PRIMITIVE)\
-                .type(xviz.PRIMITIVE_TYPES.CIRCLE)
-            builder.stream("/combine_heading")\
-                .coordinate(xviz.COORDINATE_TYPES.VEHICLE_RELATIVE)\
-                .stream_style({
-                    'stroke_width': 0.3, 
-                    'stroke_color': combine_color
-                })\
-                .category(xviz.CATEGORY.PRIMITIVE)\
-                .type(xviz.PRIMITIVE_TYPES.POLYLINE)
-            builder.stream("/combine_region")\
-                .coordinate(xviz.COORDINATE_TYPES.VEHICLE_RELATIVE)\
-                .stream_style({
-                    'stroke_width': 0.3,
-                    'stroke_color': combine_color
-                })\
-                .category(xviz.CATEGORY.PRIMITIVE)\
-                .type(xviz.PRIMITIVE_TYPES.POLYLINE)
-
-            builder.stream("/field_definition")\
-                .coordinate(xviz.COORDINATE_TYPES.IDENTITY)\
-                .stream_style({
-                    'stroke_color': [40, 150, 40, 128],
-                    'stroke_width': 0.3,
-                })\
-                .category(xviz.CATEGORY.PRIMITIVE)\
-                .type(xviz.PRIMITIVE_TYPES.POLYLINE)
-
-            # builder.stream("/predicted_path_discrete")\
-            #     .coordinate(xviz.COORDINATE_TYPES.IDENTITY)\
-            #     .stream_style({'fill_color': [0, 128, 128, 128]})\
-            #     .category(xviz.CATEGORY.PRIMITIVE)\
-            #     .type(xviz.PRIMITIVE_TYPES.CIRCLE)
-            builder.stream("/predicted_path")\
-                .coordinate(xviz.COORDINATE_TYPES.VEHICLE_RELATIVE)\
-                .stream_style({'stroke_color': [0, 128, 128, 128]})\
-                .category(xviz.CATEGORY.PRIMITIVE)\
-                .type(xviz.PRIMITIVE_TYPES.POLYLINE)
-            builder.stream("/predictive_sub_path")\
-                .coordinate(xviz.COORDINATE_TYPES.IDENTITY)\
-                .stream_style({'stroke_color': [128, 128, 0, 128]})\
-                .category(xviz.CATEGORY.PRIMITIVE)\
-                .type(xviz.PRIMITIVE_TYPES.POLYLINE)
-            # builder.stream("/commanded_path")\
-            #     .coordinate(xviz.COORDINATE_TYPES.IDENTITY)\
-            #     .stream_style({'stroke_color': [128, 0, 128, 128]})\
-            #     .category(xviz.CATEGORY.PRIMITIVE)\
-            #     .type(xviz.PRIMITIVE_TYPES.POLYLINE)
-
-            builder.stream("/planned_path")\
-                .coordinate(xviz.COORDINATE_TYPES.VEHICLE_RELATIVE)\
-                .stream_style({
-                    'stroke_width': 0.3,
-                    'stroke_color': [128, 0, 128, 128]
-                })\
-                .category(xviz.CATEGORY.PRIMITIVE)\
-                .type(xviz.PRIMITIVE_TYPES.POLYLINE)
-
-            builder.stream("/radar_fov")\
-                .coordinate(xviz.COORDINATE_TYPES.VEHICLE_RELATIVE)\
-                .stream_style({'stroke_color': [255, 0, 0, 100]})\
-                .category(xviz.CATEGORY.PRIMITIVE)\
-                .type(xviz.PRIMITIVE_TYPES.POLYLINE)
-            builder.stream("/camera_fov")\
-                .coordinate(xviz.COORDINATE_TYPES.VEHICLE_RELATIVE)\
-                .stream_style({'stroke_color': [0, 150, 200, 100]})\
-                .category(xviz.CATEGORY.PRIMITIVE)\
-                .type(xviz.PRIMITIVE_TYPES.POLYLINE)
-
-            builder.stream("/measuring_circles")\
-                .coordinate(xviz.COORDINATE_TYPES.VEHICLE_RELATIVE)\
-                .stream_style({
-                    'stroked': True,
-                    'filled': False,
-                    'stroke_width': 0.2,
-                    'stroke_color': [0, 0, 0, 20],
-                })\
-                .category(xviz.CATEGORY.PRIMITIVE)\
-                .type(xviz.PRIMITIVE_TYPES.CIRCLE)
-
-            builder.stream("/measuring_circles_lbl")\
-                .coordinate(xviz.COORDINATE_TYPES.VEHICLE_RELATIVE)\
-                .stream_style({
-                    'fill_color': [0, 0, 0]
-                })\
-                .category(xviz.CATEGORY.PRIMITIVE)\
-                .type(xviz.PRIMITIVE_TYPES.TEXT)
+            builder = get_builder()
 
             self._metadata = builder.get_message()
